// frontend/src/app/shared/types/Partner.ts
export interface Partner {
  title: string;
  cnpj: string;
  area: string;
  description: string;
  email: string;
  phone: number;
  owner: string;

  burgh: string;
  state: string;
  city: string;
  street: string;
  cep: number;
  streetNumber: number;
  complement: string;

  startDate: string;
}